--- src/main/java/com/dentallink/domain/payment/service/PaymentExternalService.java
@@ -1,9 +1,0 @@
-package com.dentallink.domain.payment.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class PaymentExternalService {
-}
