+++ src/main/java/com/dentallink/domain/payment/service/PaymentService.java
@@ -28,7 +28,7 @@ import java.nio.charset.StandardCharsets;
 
 @Service
 @RequiredArgsConstructor
-public class PaymentInternalService {
+public class PaymentService {
 
     private final PaymentRepository paymentRepository;
     private final PointAccountExternalService pointAccountExternalService;
@@ -147,57 +147,5 @@ public class PaymentInternalService {
 
     //--------------------------------- postman 테스트용-------------------
 
-    @Transactional
-    public PaymentReadyResponse testReadyPayment(Long userId, PaymentReadyRequest request) {
-        PointAccount account = pointAccountExternalService.getPointAccountByUserId(userId);
 
-        // 이미 같은 orderId가 존재하면 중복 에러 방지
-        paymentRepository.findByOrderId(request.orderId())
-                .ifPresent(p -> {
-                    throw new InvalidPaymentException(PaymentErrorCode.DUPLICATE_ORDER_ID);
-                });
-
-        Payment payment = Payment.create(account, request.amount(), request.orderId());
-        Payment saved = paymentRepository.save(payment);
-        System.out.println("결제 준비 완료: " + saved.getOrderId());
-        return PaymentReadyResponse.from(saved);
-    }
-
-    @Transactional
-    public PaymentCancelResponse testCancelPayment(String orderId) {
-        Payment payment = paymentRepository.findByOrderId(orderId)
-                .orElseThrow(() -> new InvalidPaymentException(PaymentErrorCode.NOT_FOUND_ORDER_ID));
-
-        if (payment.getStatus() != PaymentStatus.READY) {
-            throw new InvalidPaymentException(PaymentErrorCode.INVALID_PAYMENT_STATUS);
-        }
-        payment.markCancelled();
-        System.out.println("결제 취소 완료: " + orderId);
-        return PaymentCancelResponse.from(payment);
-    }
-
-
-    @Transactional
-    public PaymentResponse testConfirmPayment(Long userId, PaymentConfirmRequest request) {
-        PointAccount account = pointAccountExternalService.getPointAccountByUserId(userId);
-
-        Payment payment = paymentRepository.findByOrderId(request.orderId())
-                .orElseThrow(() -> new InvalidPaymentException(PaymentErrorCode.NOT_FOUND_ORDER_ID));
-
-        if (!payment.getAmount().equals(request.amount())) {
-            throw new InvalidPaymentException(PaymentErrorCode.INVALID_AMOUNT);
-        }
-
-        if (payment.getStatus() != PaymentStatus.READY) {
-            throw new InvalidPaymentException(PaymentErrorCode.INVALID_PAYMENT_STATUS);
-        }
-
-        // 결제 성공 처리
-        account.deposit(request.amount());
-        payment.markSuccess(request.paymentKey());
-        pointLogExternalService.createLog(account, PointLogType.DEPOSIT, request.amount());
-
-        System.out.println("결제 승인 완료: " + request.orderId());
-        return PaymentResponse.from(payment);
-    }
 }
